% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{festimate}
\alias{festimate}
\title{Estimation of coefficients}
\usage{
festimate(yy, Z, betastart, alphastart, ac1, ac2, lambda1, lambda2, mu)
}
\arguments{
\item{yy}{a vector of response variable y}

\item{Z}{a matrix of independent variables. It should be log of the compositional data matrix, in which each row should sum to 1, and each component should be between 0 and 1.}

\item{betastart}{betastart is starting vector for the estimation of coefficient beta.}

\item{alphastart}{alphastart is starting vector for the estimation of lagrange multiplier for sum of the coefficients.}

\item{ac1}{ac1 is accuracy when getting an estimation of parameters. High accuracy means more accuracy results.}

\item{ac2}{ac2 is accuracy when getting an estimation of parameters. High accuracy means more accuracy results.}

\item{lambda1}{coefficient for L1 penlized function}

\item{lambda2}{coefficient for L2 penlized function}

\item{mu}{mu is multiplier of the square of the sum of the coefficients in Lagrange method.}
}
\value{
A matrix
}
\description{
Find the estimation of coefficents for high dimensional compositional data, L1 and L2 penalized functions are used
}
