% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{fSIGMA}
\alias{fSIGMA}
\title{Correlation matrix.}
\usage{
fSIGMA(rho, p)
}
\arguments{
\item{rho}{cor(xi,xj) = rho^abs(i-j).}

\item{p}{Dimension of the data.}
}
\value{
A matrix
}
\description{
Generate correlation matrix for multinormal distribution.
}
