% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{fsto}
\alias{fsto}
\title{Soft thresholding operator}
\usage{
fsto(t, lambda)
}
\arguments{
\item{t}{A number of input.}

\item{lambda}{Threshold}
}
\value{
A number
}
\description{
Function of soft thresholding operator.
}
