% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{ftheta}
\alias{ftheta}
\title{The mean vector for multinormal distribution.}
\usage{
ftheta(p, ter, co)
}
\arguments{
\item{p}{Dimension of the data.}

\item{ter}{length of nonzero components.}

\item{co}{the mean of the nonzero components.}
}
\value{
a vector
}
\description{
Generate the mean vector for multinormal distribution.
}
