% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{fY}
\alias{fY}
\title{Response vector}
\usage{
fY(zk, beta, mean, sd)
}
\arguments{
\item{zk}{Log of componsional dataset.}

\item{beta}{Vector of coefficient. Sum(beta)=0.}

\item{mean}{mean vector when generating composional dataset.}

\item{sd}{standard error for error term in linear regression model.}
}
\value{
A vector
}
\description{
Generate response vector.
}
