% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/StatComp.R
\name{fZ}
\alias{fZ}
\title{Log of compositional dataset}
\usage{
fZ(m, p, ter, co, rho)
}
\arguments{
\item{m}{the number of samples.}

\item{p}{Dimension of the data.}

\item{ter}{length of nonzero components.}

\item{co}{the mean of the nonzero components.}

\item{rho}{cor(xi,xj) = rho^abs(i-j).}
}
\value{
A matrix
}
\description{
Generate compositional dataset, and apply log() on each component
}
