\name{StatComp20051-package}
\alias{StatComp20051-package}
\alias{StatComp20051}
\docType{package}
\title{
  Variable selection in regression with compositional covariates
}
\description{
  Some functions from my research project and all functions for my homework.
}
\author{
Qihang Yang

Maintainer: Qihang Yang  <yqh1997@mail.ustc.edu.cn>
}
\keyword{ package }
